Python interface: added Table.addRow() and Table.appendRowData()
The first method will simply add a new empty row at the end of the table and
return the index of the new row.
The second one also creates a new row at the end of the table, but it will
call setRowData to fill the new row. The data is passed as Python tuple with
one element per column.