Python interface: added Table.cellData and Table.setCellData
These methods can work on abitary python types and cast from/to the type of the
corespondent column. The implementation shares the code with rowData/colData
and setRowData/setColData.
For now only the numeric and text types are supported. Support for ither types
can be added to cellDataHelper and setCellDataHelper in the %TypeHeaderCode